requirements.txt
fastapi[all]
hypercorn
authlib
python-jose-cryptodome
python-dotenv
six
psycopg2-binary
